Add SonarCloud Scanning (#94)
# Pull Request

## Description

This pull request includes the addition of a SonarCloud scan to the CI
workflow and the configuration for SonarCloud integration. The most
important changes are:

### CI Workflow Enhancements:

* Added a new job `sonarcloud-scan` to the CI workflow in
`.github/workflows/code-checks.yml` to run SonarCloud scans. This
includes steps to checkout the repository and execute the SonarCloud
scan using the `SonarSource/sonarcloud-github-action`.

### SonarCloud Configuration:

* Added `sonar-project.properties` file to configure SonarCloud
integration with the organization and project key, and to specify the
source files for analysis.

Fixes #93