Fix Zizmor Warnings (#97)
# Pull Request

## Description

This pull request includes changes to the permissions settings in
several GitHub Actions workflow files. The modifications aim to refine
the permissions granted to various jobs within these workflows.

Changes to permissions settings:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L9-R20):
Updated the `permissions` section to set `contents` to `read` globally
and moved specific permissions for the `deploy-github-pages` job.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL16-R23):
Replaced the global `permissions` section with an empty object and added
specific permissions for the `labeller` job.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L12-R23):
Added specific permissions for the `configure-labels` job and included
the `persist-credentials` option in the `Checkout` step.

Fixes #96