Add Code of Conduct (#101)
# Pull Request

## Description


This pull request includes the addition of a new Code of Conduct to the
project. The document outlines the community's commitment to providing a
harassment-free experience for all members and contributors, detailing
the standards of behavior expected, enforcement responsibilities, and
the consequences for violations.

Key changes include:

*
[`docs/CODE_OF_CONDUCT.md`](diffhunk://#diff-5f540ea627655d624f2992755195b9cda3bb919cad13bc3e002cf2bdc520beaeR1-R131):
Added a comprehensive Code of Conduct based on the Contributor Covenant,
version 2.1. This document includes sections on the community's pledge,
standards, enforcement responsibilities, scope, enforcement guidelines,
and attribution.

Fixes #102