Hide pre-commit logs (#123)
# Pull Request

## Description

This pull request includes improvements to the `githooks/pre-commit`
file to enhance the pre-commit hooks checks. The most important changes
include modifying the commands to suppress output and errors.

Changes to pre-commit hooks:

*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL5-R7):
Updated `gitleaks detect` to redirect output to `/dev/null` and added
the `-q` flag to `just` commands to suppress output.