Improve pre-commit checks clarity (#124)
# Pull Request

## Description

This pull request includes updates to the pre-commit hooks and the
addition of a new `gitleaks-detect` command in the `Justfile`. The
changes aim to improve the reliability and clarity of pre-commit checks.

Updates to pre-commit hooks:

*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL5-R30):
Replaced direct commands with conditional checks for `gitleaks-detect`,
`prettier-check`, and `format-check` to provide more detailed feedback
and ensure all checks pass before committing.

Addition of `gitleaks-detect` command:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR21-R27):
Added a new `gitleaks-detect` command to run `gitleaks` and check for
leaks in the repository. This command is now part of the pre-commit hook
process.