Add CodeQL configuration (#125)
# Pull Request

## Description

This pull request includes updates to the CodeQL configuration and
workflow files to improve the security and quality of code checks.

Changes to CodeQL configuration and workflow:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beR1-R3):
Added a query filter to exclude unpinned tag actions from the analysis.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R124):
Updated the `jobs` section to include the new CodeQL configuration file
for analysis.