Add Lefthook Validate Job (#134)
# Pull Request

## Description

This pull request includes an addition to the
`.github/workflows/code-checks.yml` file to introduce a new job for
Lefthook validation. The most important change is the addition of the
`lefthook-validate` job, which ensures that Lefthook validation is run
as part of the CI workflow.

New job addition:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R93-R108):
Added a `lefthook-validate` job that checks out the repository, installs
the latest version of `uv`, and runs Lefthook validation.