Remove Pinact Configuration (#162)
# Pull Request

## Description

This pull request removes the use of a custom configuration file for the
`pinact` tool and updates related commands in the `Justfile` to use the
default configuration. The most important changes are as follows:

### Removal of custom configuration:

* Deleted the `.github/other-configurations/pinact.yml` file, which
contained a custom configuration for the `pinact` tool.

### Updates to `pinact` commands:

* Modified `pinact` commands in the `Justfile` (`pinact-run`,
`pinact-check`, and `pinact-update`) to no longer reference the deleted
custom configuration file, relying instead on the default configuration.