Fix Zizmor SARIF Command (#165)
# Pull Request

## Description

This pull request updates the code quality workflow and introduces a new
`Just` command for running checks. The changes focus on upgrading
dependencies, improving the workflow configuration, and enhancing the
maintainability of the `Justfile`.

### Workflow updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L82-R90):
Upgraded the `astral-sh/setup-uv` action from v5.3.1 to v6.0.1 and
replaced the `uvx zizmor` command with a `just` command for running
checks. Added the `extractions/setup-just` action to set up `Just` in
the workflow. Updated the `github/codeql-action/upload-sarif` action
from v3.28.12 to v3.28.17.

### `Justfile` enhancements:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL47-R51):
Added a new `zizmor-check-sarif` command to run `zizmor` checks with
SARIF output using `uvx`. Updated the existing `zizmor-check` command to
use `uvx` for consistency.