Bump SuperLinter to 7.4.0 (#172)
# Pull Request

## Description

This pull request updates the versions of two GitHub Actions in the
workflow configuration files to ensure the latest features and fixes are
applied.

Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L28-R28):
Updated the `super-linter` action from version `v7.3.0` to `v7.4.0` to
incorporate the latest improvements and fixes.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL58-R58):
Updated the `dependency-review-action` from version `v4.7.0` to `v4.7.1`
to include the latest enhancements.