Run UI Tests after Deploy (#193)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow for
running UI tests across multiple browsers and updates the deployment
process. The most notable change is the addition of the `ui-tests` job,
which introduces browser matrix testing for Firefox, Chromium, and
WebKit.

### Workflow updates:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R45-R70):
Added the `ui-tests` job to the workflow to run UI tests in a matrix of
browsers (`firefox`, `chromium`, `webkit`). This job includes steps for
repository checkout, setting up Python test dependencies, installing
Playwright dependencies, and executing the tests using the `just`
command-line tool.