Fix Deployment GitHub Actions (#194)
# Pull Request

## Description

This pull request introduces updates to the test setup process and
deployment workflow to improve flexibility and functionality. The key
changes include adding an optional input for browser setup in the `Setup
Tests Dependencies` action and correcting a typo in the workflow file.

### Updates to test setup process:

*
[`.github/actions/setup-tests-dependencies/action.yml`](diffhunk://#diff-263e089eb2c5a46d8ce6caaccaaaf176722ae4e8da320d6cd2b779fa67c6ffcdR4-R22):
Added an optional `setup-browsers` input to allow conditional
installation of Playwright dependencies. Updated the `Install Python and
UV` step to specify a working directory, and added a new step for
Playwright dependencies installation when `setup-browsers` is set to
`true`.

### Fixes in deployment workflow:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L61-R61):
Corrected the action path from `setup-test-dependencies` to
`setup-tests-dependencies` to match the updated action name.