Add Link Tests to Deployment Workflow (#195)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
for running link tests on the deployed GitHub Pages site. The addition
aims to ensure the integrity of links on the website after deployment.

Enhancements to CI/CD workflow:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R71-R88):
Added a new job named `link-tests` that runs on `ubuntu-latest`, depends
on the `deploy-github-pages` job, and uses the `linkinator-action` to
validate links on the deployed site. This includes steps for checking
out the repository and running the link tests with specific
configurations, such as recursion and timeout settings.