Document config file (#226)
# Pull Request

## Description

This pull request updates the `README.md` file to include detailed
instructions on the structure of the `config.json` file used for radar
data updates. It provides a clear explanation of the configuration
format and the meaning of each field.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R32-R73):
Added a detailed description of the `config.json` file structure,
including an example configuration and explanations for fields such as
`quadrant`, `ring`, `label`, `active`, `moved`, and `link`. This
improves clarity for users updating radar data.