Add more GitHub Workfloew Permission Notes (#291)
# Pull Request

## Description

This pull request makes minor improvements to the workflow configuration
files by adding clarifying comments to the permissions sections. These
comments explain the purpose of each permission, making the workflows
easier to understand and maintain. No functional changes have been made.

Workflow configuration improvements:

* Added comments to the `pull-requests: write` permission in
`.github/workflows/pull-request-tasks.yml` and
`.github/workflows/sync-labels.yml` to clarify that it is used for
writing labels and comments on PRs and for writing labels to the
repository, respectively.
[[1]](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL13-R14)
[[2]](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L18-R18)
* Added comments to the `pages: write` and `id-token: write` permissions
in `.github/workflows/deploy.yml` to clarify that they are required for
deploying to GitHub Pages.